// src/types/sanity.types.ts
export type NavLink = {
  _type: 'navLink'
  label?: string
  url?: string
  children?: Array<
    {
      _key: string
    } & NavLink
  >
}

export type NavItem = {
  _id: string
  _type: 'navItem'
  _createdAt: string
  _updatedAt: string
  _rev: string
  heading?: string
  links?: Array<
    {
      _key: string
    } & NavLink
  >
}

export type Nav = {
  _id: string
  _type: 'nav'
  _createdAt: string
  _updatedAt: string
  _rev: string
  title?: string
  navButton?: string
  items?: Array<{
    heading?: string
    links?: Array<
      {
        _key: string
      } & NavLink
    >
    _type: 'navItem'
    _key: string
  }>
}